# specs/001-wiki-platform/contracts/api.yaml
openapi: 3.0.3
info:
  title: Wiki Platform API
  description: API for the wiki platform with article management, user profiles, and social features
  version: 1.0.0
  contact:
    name: Wiki Platform Support
    email: support@wikiplatform.example.com
servers:
  - url: https://api.wikipatform.example.com/v1
    description: Production server
  - url: https://staging-api.wikipatform.example.com/v1
    description: Staging server
paths:
...
  /articles:
    post:
      summary: Create article
      description: Create a new article draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreate'
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List articles
      description: Retrieve a paginated list of articles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArticleResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /articles/{articleId}:
    get:
      summary: Get article
      description: Retrieve a specific article by ID
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update article
      description: Update an existing article
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpdate'
      responses:
        '200':
          description: Article updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the article owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete article
      description: Delete an article (soft delete - moves to archived status)
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Article deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the article owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /articles/{articleId}/publish:
    post:
      summary: Publish article
      description: Publish a draft article to make it publicly visible
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          description: Article cannot be published (e.g., no content)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the article owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /articles/{articleId}/revisions:
    get:
      summary: Get article revisions
      description: Retrieve the revision history for an article
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article revisions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  revisions:
                    type: array
                    items:
                      $ref: '#/components/schemas/RevisionResponse'
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        role:
          type: string
          enum: [user, moderator, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        contributionCount:
          type: integer
...
    ArticleCreate:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          minLength: 10
        summary:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [draft, published]
          default: draft
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
        tagNames:
          type: array
          items:
            type: string
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          minLength: 10
        summary:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [draft, published, archived]
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
        tagNames:
          type: array
          items:
            type: string
    ArticleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        summary:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        author:
          $ref: '#/components/schemas/UserResponse'
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        viewCount:
          type: integer
        isFeatured:
          type: boolean
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        parentId:
          type: string
          format: uuid
        articleCount:
          type: integer
    TagResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        usageCount:
          type: integer
    RevisionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/UserResponse'
        title:
          type: string
        content:
          type: string
        changeSummary:
          type: string
        revisionNumber:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object